refactor: move to bitcoinerlab.js
